Nest the adapters inside TypeDeclarationSyntaxExtensions
